+++ saga-gis/src/modules/terrain_analysis/ta_hydrology/Flow_RecursiveUp.cpp
@@ -181,12 +181,12 @@ void CFlow_RecursiveUp::On_Create(void)
 	//-----------------------------------------------------
 	On_Destroy();
 
-	Flow	= (double ***)SG_Malloc(    Get_NY()  * sizeof(double **));
-	p		= (double   *)SG_Malloc((long) 8 * Get_NCells() * sizeof(double   ));
+	Flow	= (double ***)SG_Malloc(    Get_NY    () * sizeof(double **));
+	p		= (double   *)SG_Malloc(8 * Get_NCells() * sizeof(double   ));
 
 	for(y=0; y<Get_NY(); y++)
 	{
-		Flow[y]	= (double **)SG_Malloc( Get_NX()  * sizeof(double  *));
+		Flow[y]	= (double **)SG_Malloc( Get_NX    () * sizeof(double  *));
 
 		for(x=0; x<Get_NX(); x++, p+=8)
 		{
@@ -199,7 +199,7 @@ void CFlow_RecursiveUp::On_Create(void)
 
 	Method	= Parameters("Method")->asInt();
 
-	memset(Flow[0][0], 0, (long) 8 * Get_NCells() * sizeof(double) );
+	memset(Flow[0][0], 0, 8 * Get_NCells() * sizeof(double) );
 
 	for(y=0; y<Get_NY(); y++)
 	{
